Migrate TOML-util-related functions from `FuzzerConfig` to `TOMLUtils`
