+++ core/program_info.cpp
@@ -1,6 +1,7 @@
 #include "backends/p4tools/modules/rtsmith/core/program_info.h"
 
 #include "backends/p4tools/common/compiler/compiler_target.h"
+#include "backends/p4tools/modules/rtsmith/core/toml_utils.h"
 
 namespace P4::P4Tools::RtSmith {
 
@@ -20,11 +21,11 @@ const ::p4::config::v1::P4Info *ProgramInfo::getP4Info() const { return p4runtim
 const FuzzerConfig &ProgramInfo::getFuzzerConfig() const { return _fuzzerConfig; }
 
 void ProgramInfo::loadFuzzerConfig(std::filesystem::path path) {
-    _fuzzerConfig.overrideFuzzerConfigs(path);
+    TOMLUtils::overrideFuzzerConfigs(_fuzzerConfig, path);
 }
 
 void ProgramInfo::loadFuzzerConfigInString(std::string configInString) {
-    _fuzzerConfig.overrideFuzzerConfigsInString(configInString);
+    TOMLUtils::overrideFuzzerConfigsInString(_fuzzerConfig, configInString);
 }
 
 }  // namespace P4::P4Tools::RtSmith
